优化解析函数的文件结构（基础节点）
